Cleaned up "Add feed" and "Edit feed" wizard
The simple dialog provides a better overview
of the entered data, and allows the user to
use less touches to add a feed. It's possible
to simply use the keyboard ("Enter" goes to
the next field and adds the feed when pressed
in the URL field) for the complete interaction.